Tweaked spannws on Stage 2, and tweaked tilmap position to allow halves.

+++ src/Source/Code/CorePlugin/World/Level/TilemapObjectPositioner.cs
@@ -14,8 +14,8 @@ namespace Khronos.World.Level
     public class TilemapObjectPositioner : Component, ICmpInitializable
     {
 
-        public int OffsetX { get; set; }
-        public int OffsetY { get; set; }
+        public float OffsetX { get; set; }
+        public float OffsetY { get; set; }
 
         public virtual void OnInit(InitContext context)
         {
